Refactor BookingStore and update types; enhance image upload handling in BookingModal
