+++ src/store/types.ts
@@ -1,10 +1,10 @@
 import { ReactNode } from 'react'
-import { BookingStore } from '~/container/MentorProfile/components/Booking/BookingStore/types'
+import { BookingStore } from '~/container/MentorProfile/components/Booking/BookingStore/BookingStore'
 
 export interface Props {
     children: ReactNode
 }
 
-export interface RootStoreAPI {
+export interface RootStore {
     bookingStore: BookingStore
 }
